+++ node/mock-task.ts
@@ -74,8 +74,13 @@ export function setResourcePath(path: string): void {
     // nothing in mock
 }
 
-export function loc(key: string): string {
-    return 'loc_mock_' + key;
+export function loc(key: string, ...args: any[]): string {
+    let str: string = 'loc_mock_' + key;
+    if (args.length) {
+        str += ' ' + args.join(' ');
+    }
+
+    return str;
 }
 
 //-----------------------------------------------------
@@ -201,12 +206,12 @@ export class FsStats implements fs.Stats {
 
 export function stats(path: string): FsStats {
     var fsStats = new FsStats();
-    fsStats.setAnswers(mock.getResponse('stats', path) || {});
+    fsStats.setAnswers(mock.getResponse('stats', path, module.exports.debug) || {});
     return fsStats;
 }
 
 export function exist(path: string): boolean {
-    return mock.getResponse('exist', path) || false;
+    return mock.getResponse('exist', path, module.exports.debug) || false;
 }
 
 export interface FsOptions {
@@ -220,11 +225,11 @@ export function writeFile(file: string, data: string|Buffer, options?: string|Fs
 }
 
 export function osType(): string {
-    return mock.getResponse('osType', 'osType');
+    return mock.getResponse('osType', 'osType', module.exports.debug);
 }
 
 export function cwd(): string {
-    return mock.getResponse('cwd', 'cwd');
+    return mock.getResponse('cwd', 'cwd', module.exports.debug);
 }
 
 //-----------------------------------------------------
@@ -253,7 +258,7 @@ export function popd(): void {
 
 export function checkPath(p: string, name: string): void {
     module.exports.debug('check path : ' + p);
-    if (!p || !mock.getResponse('checkPath', p)) {
+    if (!p || !mock.getResponse('checkPath', p, module.exports.debug)) {
         throw new Error('Not found ' + p);
     }
 }
@@ -278,7 +283,7 @@ export function resolve(): string {
 }
 
 export function which(tool: string, check?: boolean): string {
-    var response = mock.getResponse('which', tool);
+    var response = mock.getResponse('which', tool, module.exports.debug);
     if (check) {
         checkPath(response, tool);
     }
@@ -286,7 +291,7 @@ export function which(tool: string, check?: boolean): string {
 }
 
 export function ls(options: string, paths: string[]): string[] {
-    var response = mock.getResponse('ls', paths[0]);
+    var response = mock.getResponse('ls', paths[0], module.exports.debug);
     if(!response){
         return [];
     }
@@ -294,17 +299,17 @@ export function ls(options: string, paths: string[]): string[] {
 }
 
 export function cp(source: string, dest: string): void {
-    console.log('###copying###');
+    module.exports.debug('###copying###');
     module.exports.debug('copying ' + source + ' to ' + dest);
 }
 
 export function find(findPath: string): string[] {
-    return mock.getResponse('find', findPath);
+    return mock.getResponse('find', findPath, module.exports.debug);
 }
 
 export function rmRF(path: string): void {
     module.exports.debug('rmRF ' + path);
-    var response = mock.getResponse('rmRF', path);
+    var response = mock.getResponse('rmRF', path, module.exports.debug);
     if (!response['success']) {
         module.exports.setResult(1, response['message']);
     }
@@ -313,40 +318,6 @@ export function rmRF(path: string): void {
 export function mv(source: string, dest: string, force: boolean, continueOnError?: boolean): boolean {
     module.exports.debug('moving ' + source + ' to ' + dest);
     return true;
-}
-
-export function glob(pattern: string): string[] {
-    module.exports.debug('glob ' + pattern);
-
-    var matches: string[] = mock.getResponse('glob', pattern);
-    module.exports.debug('found ' + matches.length + ' matches');
-
-    if (matches.length > 0) {
-        var m = Math.min(matches.length, 10);
-        module.exports.debug('matches:');
-        if (m == 10) {
-            module.exports.debug('listing first 10 matches as samples');
-        }
-
-        for (var i = 0; i < m; i++) {
-            module.exports.debug(matches[i]);
-        }
-    }
-
-    return matches;
-}
-
-export function globFirst(pattern: string): string {
-    module.exports.debug('globFirst ' + pattern);
-    var matches = glob(pattern);
-
-    if (matches.length > 1) {
-        module.exports.warning('multiple workspace matches.  using first.');
-    }
-
-    module.exports.debug('found ' + matches.length + ' matches');
-
-    return matches[0];
 }
 
 //-----------------------------------------------------
@@ -383,6 +354,9 @@ export function tool(tool: string): trm.ToolRunner {
 //-----------------------------------------------------
 // Matching helpers
 //-----------------------------------------------------
+module.exports.filter = task.filter;
+module.exports.match = task.match;
+
 // redefine to avoid folks having to typings install minimatch
 export interface MatchOptions {
     debug?: boolean;
@@ -397,21 +371,10 @@ export interface MatchOptions {
     nonegate?: boolean;
     flipNegate?: boolean;
 }
-export function match(list: string[], pattern: string, options?: MatchOptions): string[];
-export function match(list: string[], patterns: string[], options?: MatchOptions): string[];
-export function match(list: string[], pattern: any, options?: MatchOptions): string[] {
-    return mock.getResponse('match', pattern) || [];
-}
 
-export function matchFile(list, pattern, options): string[] {
-    return mock.getResponse('match', pattern) || [];
-}
-
-export function filter(pattern, options): any {
-    var filterList = mock.getResponse('filter', pattern) || [];
-	return function(pattern, i, arr) {
-		return filterList.indexOf(pattern) >= 0;
-	}
+export function findMatch(defaultRoot: string, patterns: string[] | string) : string[] {
+    let responseKey: string = typeof patterns == 'object' ? (patterns as string[]).join('\n') : patterns as string;
+    return mock.getResponse('findMatch', responseKey, module.exports.debug);
 }
 
 //-----------------------------------------------------